Fiddled with numbers

+++ spg/src/com/spg/chapter1/shields/Shield.java
@@ -2,5 +2,5 @@ package com.spg.chapter1.shields;
 
 
 public interface Shield {
-	public int getStrength();
+	public double getStrength();
 }
